fix: change navigation links
